Add cleanup after a player has been deleted and fix get_tile
Now the tiles have no owner or occupant if there was a city or unit owned by the deleted player. Nothing is blocked when it comes to moving and area assigning.
Getting a tile worked for negative coords because Python is too powerful. Now it's fixed and the Earth is flat again!